requirements.txt
django
django-crispy-forms
django-environ
graphene 
graphene-django
pytest
django-multiselectfield